refactor: move season players slice into core application
