+++ backend/src/api/dependencies/use_cases.py
@@ -31,6 +31,9 @@ from core.application.use_cases.manage_pena_seasons_usecase import (
 from core.application.use_cases.manage_season_lifecycle_usecase import (
     ManageSeasonLifecycleUseCase,
 )
+from core.application.use_cases.manage_season_players_usecase import (
+    ManageSeasonPlayersUseCase,
+)
 from core.application.use_cases.register_admin_usecase import RegisterAdminUseCase
 from core.application.use_cases.register_user_usecase import RegisterUserUseCase
 from core.application.use_cases.update_player_profile_usecase import (
@@ -42,9 +45,6 @@ from persistence.application.use_cases.manage_season_competition_usecase import 
 )
 from persistence.application.use_cases.manage_season_matches_usecase import (
     ManageSeasonMatchesUseCase,
-)
-from persistence.application.use_cases.manage_season_players_usecase import (
-    ManageSeasonPlayersUseCase,
 )
 from persistence.infrastructure.repository.db.nationality_query_repository import (
     SqlAlchemyNationalityQueryRepository,
