+++ backend/src/core/application/ports/__init__.py
@@ -11,6 +11,7 @@ from .player_profile_port import PlayerProfilePort
 from .registration_port import AdminRegistrationPort, UserRegistrationPort
 from .season_competition_port import SeasonCompetitionPort
 from .season_match_insights_port import SeasonMatchInsightsPort
+from .season_player_port import SeasonPlayerPort
 
 __all__ = [
     "AdminRegistrationPort",
@@ -26,5 +27,6 @@ __all__ = [
     "PlayerProfilePort",
     "SeasonCompetitionPort",
     "SeasonMatchInsightsPort",
+    "SeasonPlayerPort",
     "UserRegistrationPort",
 ]
